+++ cf/read_write/um/functions.py
@@ -1,28 +1,30 @@
 from ..constants import _file_to_fh
-from ..functions import open_files_threshold_exceeded, close_one_file #, abspath
+from ..functions import (open_files_threshold_exceeded,  # abspath
+                         close_one_file)
 
 from .umread.umfile import File, UMFileException
 
 _file_to_UM = _file_to_fh.setdefault('UM', {})
 
+
 def _open_um_file(filename, aggregate=True, fmt=None, word_size=None,
                   byte_ordering=None):
     '''Open a UM fields file or PP file and read it into a
     `umread.umfile.File` object.
-    
+
     If there is already a `umread.umfile.File` object for the file
     then it is returned with an open file descriptor.
-    
+
     :Parameters:
-    
+
         filename: `str`
             The file to be opened.
-    
+
     :Returns:
-    
+
         `umread.umfile.File`
             The opened file with an open file descriptor.
-    
+
     **Examples:**
 
     '''
@@ -65,20 +67,21 @@ def _open_um_file(filename, aggregate=True, fmt=None, word_size=None,
 
     return f
 
+
 def _close_um_file(filename):
     '''Close a PP or UM fields file.
 
     Does nothing if the file is already closed.
-    
+
     :Parameters:
-    
+
         filename: `str`
             The file to be closed.
-    
+
     :Returns:
-    
+
         `None`
-    
+
     **Examples:**
 
     '''