+++ Changelog.rst
@@ -1,3 +1,21 @@
+version 3.3.0
+-------------
+----
+
+**2020-??-??**
+
+* Changed the API to `cf.Field.convolution_filter`: renamed the
+  ``weights`` parameter to ``window``.
+* Reinstated `True` as a permitted value of the ``weights`` keyword of
+  `cf.Field.collapse`.
+* New method: `cf.Field.moving_window`
+  (https://github.com/NCAS-CMS/cf-python/issues/44)
+* New method: `cf.Data.convolution_filter`
+* New keyword parameter to `cf.Field.weights`: ``axes``
+* New permitted values to ``coordinate`` keyword parameter of
+  `cf.Field.collapse` and `cf.Field.cumsum`: ``'minimum'``,
+  ``'maximum'``
+
 version 3.2.0
 -------------
 ----
@@ -6,6 +24,8 @@ version 3.2.0
 
 * First release for CF-1.8 (does not include netCDF hierarchical
   groups functionality).
+* Deprecated `True` as a permitted value of the ``weights`` keyword of
+  `cf.Field.collapse`.
 * New methods: `cf.Data.compressed`, `cf.Data.diff`
 * New function: `cf.implementation`
 * New methods completing coverage of the inverse trigonometric and
@@ -37,7 +57,7 @@ version 3.2.0
   ``True`` where it would always error before returning dictionary of methods.
 * Fixed bug in `cf.Data.where` that meant the units were not taken into
   account when the condition was a `cf.Query` object with specified units.
-* Addressed many 'TO DO' placeholders in the documentation.
+* Addressed many 'TODO' placeholders in the documentation.
 
 version 3.1.0
 -------------
@@ -335,7 +355,7 @@ version 2.2.5
 -------------
 ----
 
-**2018-7-02**
+**2018-07-02**
 
 * Fixed bug with HDF chunk sizes that prevented the writing of large
   files
@@ -508,6 +528,12 @@ version 2.0.4
 * Bug fix in `cf.Field.where`
 * Bug fix when initializing a field with the source parameter
 * Changed default output format to NETCDF4 (from NETCDF3_CLASSIC)
+
+version 2.0.3
+-------------
+----
+
+**2017-08-01**
 
 version 2.0.1.post1
 -------------------